Hardcode knowlegdge about compability aliases for module names in the email package.
In python 2.5 the email package moved to PEP8 compliant names for modules, but
kept the old names alive through some code in email.__init__. This patches
teaches modulegraph about that code.    The compatibility aliases are not present
in python 3.x.

With this patch it is possible to remove a heavy-handed recipe from py2app.