Provide useful dependency information on the edges in the module graph.
Note: this code is not yet battle-hardened and hence is likely to contain bugs.

This patch makes most of the API related to loading modules private (methods such as 'load_tail'
are renamed to names with a leading underscore).

More importantly, the edges in the graph are now annotated with an instance of "DependencyInfo" that
contains information about the location of the import statement(s) that caused this edge to exist:
is it in global code or a function body, is the import optional (in an "if" statement
or "try-except" block), and is the import on the RHS of a "from" import?

The primary use-case of this new information is to make it possible to have better reporting of
modules that seem to be missing: missing modules that are unconditionally imported tent to be more
important than missing modules that are optionally imported.