Attempt at reducing the maximum amount of recursion needed to build a dependency graph.
Before this patch the AST walker, which is a recursive process, called back into
the modulegraph builder which is also recursive. The new code collects the list of
imports in the AST walker and afterwards processes them.

This should make it more likely that the dependency graph can be created without
running into the recursion limit.

See also #37